+++ CHANGELOG.md
@@ -1,3 +1,4 @@
 - 17 Aug: cambiato il senso di gioco da orario ad antiorario, centralizzata la logica dell'ordine dentro GameState
 - 17 Aug: conversione del progetto da js puro a Typescript
 - 18 Aug: centralizzato il seme delle carte nell'oggetto Suit / Suits
+- 19 Aug: Creato Event Bus e sound Manager, aggiunto audio per il gioco delle carte e della presa.